return the latest apiversion in Content-Type
when a client requests an api version that cannot be served, send
back the latest available apiversion with the Content-Type header,
not the one that the client has specified

If the client would request e.g. for 4.2 it doesn't make sense to
return him the wrong unservable version, but instead the one that
can be served.

this also fixes the out of scope variables major and minor