perform the apache switch within the database api
on a fresh installation either database/new or database/connect are
being used.
to avoid having to do another api call after setting up the database
to initialize crowbar, it is more efficient to do that within the
existing database api's.
it also brings consistency to the way the upgrade related database
endpoints are working, as they do exactly the same (plus migrations)